test: gate integration tests behind build tag and align CI verification
Move store and vendor live-API tests behind //go:build integration,
split make test from make test-integration, and run the full suite in
CI with -tags=integration and PostgreSQL. Also reject zero release IDs
in REST/GraphQL and document the ids filter in api_design.md.
