+++ AGENTS.md
@@ -126,13 +126,19 @@ Primary verification command:
 make check
 ```
 
-This is the canonical local verification entrypoint. It runs, in order: `imports` (`goimports`), `fmt-check` (`gofmt -s -l`, matching CI’s go fmt check), `lint-local` (full-repo `golangci-lint` with CGO enabled), and `test` (`CGO_ENABLED=1` `go test -cover ./...`). Run `make fmt` to apply `gofmt -s -w` fixes when `fmt-check` fails.
+This is the canonical local verification entrypoint. It runs, in order: `imports` (`goimports`), `fmt-check` (`gofmt -s -l`, matching CI's go fmt check), `lint-local` (full-repo `golangci-lint` with CGO enabled), `test` (unit tests, `CGO_ENABLED=1 go test -cover ./...`), and `test-integration` (full suite, `CGO_ENABLED=1 go test -tags=integration -cover ./...`). Run `make fmt` to apply `gofmt -s -w` fixes when `fmt-check` fails.
+
+**Tests are split by build tag:**
+- **Unit tests** (`make test`): no build tag; no external dependencies for most packages.
+- **Integration tests** (`//go:build integration`): store tests require PostgreSQL (testcontainers when `TEST_DB_HOST` is unset, or `TEST_DB_*` for an external DB); vendor `client_integration_test.go` files call live external APIs.
+
+**Note:** `-tags=integration` is additive — `make test-integration` runs unit tests and integration-tagged tests together (same as CI).
 
 The lint profile enforces cyclomatic and cognitive complexity, function and file length, and doc quality expectations for the code it analyzes.
 
 **Note on lightweight mode:** The repository supports a CGO-disabled lightweight Docker mode for deployment (`make build`, default `quickstart`). Canonical verification (`make check` and CI) uses `CGO_ENABLED=1` tests and, in CI, a full-media Docker build (`CGO_ENABLED=1`). Run `make test-lightweight-build` manually when you need to verify the lightweight binary and stub media path; it is not part of `make check` or CI.
 
-CI still defines its own exact steps in `.github/workflows/test.yaml` and `.github/workflows/lint.yaml`. Run additional targeted generation or build checks when relevant. If you cannot run the full expected verification, say so explicitly.
+CI defines its own exact steps in `.github/workflows/test.yaml` (full test suite with `-tags=integration` and a PostgreSQL service) and `.github/workflows/lint.yaml`. Run additional targeted generation or build checks when relevant. If you cannot run the full expected verification, say so explicitly.
 
 ## Review and done
