+++ docs/api_design.md
@@ -79,12 +79,13 @@ Design rules:
 
 ### List releases (`GET /api/v1/releases`)
 
-- Returns a paginated list of release metadata rows filtered by **`vendor`** and/or **`vendor_release_id`**. At least one filter is required; when both are provided they are combined with AND semantics.
+- Returns a paginated list of release metadata rows filtered by **`ids`**, **`vendor`**, and/or **`vendor_release_id`**. At least one filter is required; when multiple are provided they are combined with AND semantics.
+- **`ids`:** One or more internal release IDs (positive integers). Repeat the parameter for multiple values: `?ids=1&ids=2`. All IDs must be positive; zero is rejected.
 - **`vendor`:** `artblocks`, `feralfile`, `fxhash`, or `objkt` (all vendors that populate the `releases` table).
 - **`vendor_release_id`:** External release key as stored during indexing (FF series UUID, AB `{chainID}-{contract}-{projectID}`, fxhash generative token numeric id, or objkt custom-collection KT1 contract address). May be used without `vendor`; results can span vendors in theory but are typically 0–1 rows.
 - **Response:** `items` array of release metadata (`id`, `vendor`, `vendor_release_id`, optional `name`, optional `total_mints`). **No `members`** — use `GET /api/v1/releases/{id}` or `GET /api/v1/tokens?release_id=...` for member tokens.
 - **Pagination:** `limit` (default **20**, max **255**), `offset` (default **0**); empty match set returns **`200`** with `"items": []`.
-- **GraphQL:** `releases(vendor, vendor_release_id, limit, offset)` exposes the same contract via `ReleaseList`.
+- **GraphQL:** `releases(ids, vendor, vendor_release_id, limit, offset)` exposes the same contract via `ReleaseList`.
 
 ### Release endpoint (`GET /api/v1/releases/{id}`)
