+++ internal/store/pg_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package store
 
 import (
