+++ internal/store/store_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package store
 
 import (
